+++ CHANGELOG.md
@@ -21,6 +21,9 @@ a release.
 ### Added
 - Tree: `setSibling()` and `getSibling()` methods in the `Node` interface through the BC `@method` annotation
 
+### Changed
+- Removed conflict against "doctrine/cache" < 1.11, as this library is not used.
+
 ### Fixed
 - Tree: Creation of dynamic `Node::$sibling` property, which is deprecated as of PHP >= 8.2